apps/contract-generator/requirements.txt
flask==3.0.0
anthropic==0.42.0
stripe==7.0.0
reportlab==4.0.7
python-dotenv==1.0.0
gunicorn==21.2.0